+++ app/Domains/Auth/Http/Controllers/Backend/User/DeactivatedUserController.php
@@ -4,13 +4,12 @@ namespace App\Domains\Auth\Http\Controllers\Backend\User;
 
 use App\Domains\Auth\Models\User;
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 /**
  * Class UserStatusController.
  */
-class DeactivatedUserController extends Controller
+class DeactivatedUserController
 {
     /**
      * @var UserService