+++ app/Domains/Auth/Http/Controllers/Backend/User/DeletedUserController.php
@@ -4,12 +4,11 @@ namespace App\Domains\Auth\Http\Controllers\Backend\User;
 
 use App\Domains\Auth\Models\User;
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 
 /**
  * Class DeletedUserController.
  */
-class DeletedUserController extends Controller
+class DeletedUserController
 {
     /**
      * @var UserService