+++ app/Domains/Auth/Http/Controllers/Backend/User/UserPasswordController.php
@@ -6,12 +6,11 @@ use App\Domains\Auth\Http\Requests\Backend\User\EditUserPasswordRequest;
 use App\Domains\Auth\Http\Requests\Backend\User\UpdateUserPasswordRequest;
 use App\Domains\Auth\Models\User;
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 
 /**
  * Class UserPasswordController.
  */
-class UserPasswordController extends Controller
+class UserPasswordController
 {
     /**
      * @var UserService