+++ app/Domains/Auth/Http/Controllers/Frontend/Auth/ResetPasswordController.php
@@ -3,7 +3,6 @@
 namespace App\Domains\Auth\Http\Controllers\Frontend\Auth;
 
 use App\Domains\Auth\Rules\UnusedPassword;
-use App\Http\Controllers\Controller;
 use App\Providers\RouteServiceProvider;
 use Illuminate\Foundation\Auth\ResetsPasswords;
 use Illuminate\Http\Request;
@@ -12,7 +11,7 @@ use LangleyFoxall\LaravelNISTPasswordRules\PasswordRules;
 /**
  * Class ResetPasswordController.
  */
-class ResetPasswordController extends Controller
+class ResetPasswordController
 {
     /*
     |--------------------------------------------------------------------------