+++ app/Http/Controllers/Backend/DashboardController.php
@@ -2,15 +2,13 @@
 
 namespace App\Http\Controllers\Backend;
 
-use App\Http\Controllers\Controller;
-
 /**
  * Class DashboardController.
  */
-class DashboardController extends Controller
+class DashboardController
 {
     /**
-     * @return \Illuminate\View\View
+     * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
      */
     public function index()
     {