+++ app/Http/Controllers/Frontend/User/ProfileController.php
@@ -3,13 +3,12 @@
 namespace App\Http\Controllers\Frontend\User;
 
 use App\Domains\Auth\Services\UserService;
-use App\Http\Controllers\Controller;
 use App\Http\Requests\Frontend\User\UpdateProfileRequest;
 
 /**
  * Class ProfileController.
  */
-class ProfileController extends Controller
+class ProfileController
 {
     /**
      * @param  UpdateProfileRequest  $request