+++ app/Http/Controllers/LocaleController.php
@@ -5,7 +5,7 @@ namespace App\Http\Controllers;
 /**
  * Class LocaleController.
  */
-class LocaleController extends Controller
+class LocaleController
 {
     /**
      * @param $locale