+++ config/boilerplate.php
@@ -16,9 +16,9 @@ return [
     'access' => [
         'captcha' => [
             'configs' => [
-                'site_key'   => env('INVISIBLE_RECAPTCHA_SITEKEY'),
+                'site_key' => env('INVISIBLE_RECAPTCHA_SITEKEY'),
                 'secret_key' => env('INVISIBLE_RECAPTCHA_SECRETKEY'),
-                'options'   => [
+                'options' => [
                     'hidden' => false,
                     'location' => 'bottomright',
                     'timeout' => 5,
@@ -124,15 +124,6 @@ return [
 
     /*
     |--------------------------------------------------------------------------
-    | Google Analytics
-    |--------------------------------------------------------------------------
-    |
-    | Found in views/includes/partials/ga.blade.php
-    */
-    'google_analytics' => env('GOOGLE_ANALYTICS', 'UA-XXXXX-X'),
-
-    /*
-    |--------------------------------------------------------------------------
     | Locale
     |--------------------------------------------------------------------------
     |
@@ -176,6 +167,8 @@ return [
             'no' => ['name' => 'Norwegian', 'rtl' => false],
             'pl' => ['name' => 'Polish', 'rtl' => false],
             'pt_BR' => ['name' => 'Brazilian Portuguese', 'rtl' => false],
+            'pt_PT' => ['name' => 'Portuguese', 'rtl' => false],
+            'ro' => ['name' => 'Romana', 'rtl' => false],
             'ru' => ['name' => 'Russian', 'rtl' => false],
             'sv' => ['name' => 'Swedish', 'rtl' => false],
             'th' => ['name' => 'Thai', 'rtl' => false],