+++ resources/js/frontend/app.js
@@ -7,7 +7,7 @@
 require('../bootstrap');
 require('../plugins');
 
-window.Vue = require('vue');
+import Vue from 'vue';
 
 /**
  * The following block of code may be used to automatically register your