+++ resources/lang/pt_PT/auth.php
@@ -1,0 +1,19 @@
+<?php
+
+return [
+    /*
+    |--------------------------------------------------------------------------
+    | Authentication Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are used during authentication for various
+    | messages that we need to display to the user. You are free to modify
+    | these language lines according to your application's requirements.
+    |
+    |--------------------------------------------------------------------------
+    */
+
+    'failed' => 'Essas credenciais não foram encontradas na nossa base de dados.',
+    'password' => 'A palavra-passe introduzida está incorreta.',
+    'throttle' => 'Excedeu as tentativas de login. Por favor tente novamente em :seconds segundos.',
+];