+++ resources/lang/ro/passwords.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'reset' => 'Parola dumneavoastra a fost resetata!',
+    'sent' => 'V-am trimis un email cu linkul de resetare al parolei!',
+    'throttled' => 'Va rugam sa asteptati inainte de a incerca din nou.',
+    'token' => 'Token-ul de resetare al parolei este invalid.',
+    'user' => 'Nu putem gasi un user cu aceasta adresa de e-mail.',
+
+];